+++ Plugins/iOS/SentryObjC.xcframework~/ios-arm64/SentryObjC.framework/Headers/SentryObjCOptions.h
@@ -4,10 +4,12 @@
 #    import "SentryObjCDefines.h"
 #    import "SentryObjCLastRunStatus.h"
 #    import "SentryObjCLevel.h"
+#    import "SentryObjCTransaction.h"
 #else
 #    import <SentryObjC/SentryObjCDefines.h>
 #    import <SentryObjC/SentryObjCLastRunStatus.h>
 #    import <SentryObjC/SentryObjCLevel.h>
+#    import <SentryObjC/SentryObjCTransaction.h>
 #endif
 
 @class SentryObjCBreadcrumb;
@@ -43,10 +45,17 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @property (nonatomic) BOOL debug;
 
+#if SDK_V10
+/**
+ * Minimum log level to be used if debug is enabled.
+ * @note Default is @c SentryObjCLevelWarning.
+ */
+#else
 /**
  * Minimum log level to be used if debug is enabled.
  * @note Default is @c SentryObjCLevelDebug.
  */
+#endif // SDK_V10
 @property (nonatomic) SentryObjCLevel diagnosticLevel;
 
 /**
@@ -88,6 +97,18 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic) BOOL enableCrashHandler;
 
 /**
+ * When enabled, the SDK introspects memory contents during a crash.
+ * Any Objective-C objects or C strings near the stack pointer or referenced by
+ * CPU registers or exceptions will be recorded in the crash report, along with
+ * their contents. This can help diagnose crashes, but may include sensitive data in memory.
+ *
+ * Disabling this will also disable the automatic 'memory corruption' diagnosis for crashes.
+ *
+ * @note Default value is @c NO.
+ */
+@property (nonatomic) BOOL enableMemoryIntrospection;
+
+/**
  * How many breadcrumbs do you want to keep in memory?
  * @note Default is 100.
  */
@@ -109,18 +130,27 @@ NS_ASSUME_NONNULL_BEGIN
 /// This block can be used to modify the event before it will be serialized and sent.
 @property (nonatomic, copy, nullable) SentryObjCEvent *_Nullable (^beforeSend)(SentryObjCEvent *);
 
+#if SDK_V10
+/// This block can be used to modify a transaction before it will be serialized and sent.
+@property (nonatomic, copy, nullable) SentryObjCTransaction *_Nullable (^beforeSendTransaction)
+    (SentryObjCTransaction *);
+#endif // SDK_V10
+
 /**
  * Use this callback to drop or modify a span before the SDK sends it to Sentry.
  * Return @c nil to drop the span.
  */
 @property (nonatomic, copy, nullable) SentryObjCSpan *_Nullable (^beforeSendSpan)(SentryObjCSpan *);
 
+#if !SDK_V10
 /**
  * When enabled, the SDK sends logs to Sentry. Logs can be captured using the
  * @c SentryObjCSDK.logger API, which provides structured logging with attributes.
  * @note Default value is @c NO.
+ * @note In v10 and later, logs are always enabled. Remove this option when upgrading.
  */
 @property (nonatomic) BOOL enableLogs;
+#endif // !SDK_V10
 
 /// This block can be used to modify the breadcrumb before it will be serialized and sent.
 @property (nonatomic, copy, nullable) SentryObjCBreadcrumb *_Nullable (^beforeBreadcrumb)
@@ -233,11 +263,13 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @property (nonatomic) NSUInteger maxAttachmentSize;
 
+#if !SDK_V10
 /**
  * When enabled, the SDK sends personal identifiable information along with events.
  * @note The default is @c NO.
  */
 @property (nonatomic) BOOL sendDefaultPii;
+#endif // !SDK_V10
 
 /**
  * When enabled, the SDK tracks performance for UIViewController subclasses and HTTP requests
@@ -557,6 +589,15 @@ NS_ASSUME_NONNULL_BEGIN
  * @note Default value is @c YES.
  */
 @property (nonatomic) BOOL enablePreWarmedAppStartTracing;
+
+#    if !SDK_V10
+/**
+ * When enabled, the SDK sends a standalone app start transaction instead of attaching app
+ * start data to the first UIViewController transaction.
+ * @note Default value is @c NO.
+ */
+@property (nonatomic) BOOL enableStandaloneAppStartTracing;
+#    endif // !SDK_V10
 
 /**
  * When enabled, the SDK reports non-fully-blocking app hangs. A non-fully-blocking app hang is